+++ apps/widget-configurator/src/hooks/useWidgetParamsAndSettings.ts
@@ -175,7 +175,7 @@ function buildWidgetParams(configuratorState: ConfiguratorState | null): CowSwap
     // Layout:
 
     disableScrollbars,
-    iframeStyle,
+    rootStyle,
     bodyWrapperStyle,
     cardStyle,
 
@@ -185,6 +185,7 @@ function buildWidgetParams(configuratorState: ConfiguratorState | null): CowSwap
     disableProgressBar,
     disablePostTradeTips,
     disableTokenImport,
+    disableSwitchingTokens,
     hideRecentTokens,
     hideFavoriteTokens,
     hideBridgeInfo,
@@ -216,7 +217,7 @@ function buildWidgetParams(configuratorState: ConfiguratorState | null): CowSwap
   } = configuratorState
 
   const baseUrl = rawBaseUrl || CONFIGURATOR_DEFAULT_WIDGET_BASE_URL
-  const legacyIframeDimensionParams = getLegacyIframeDimensionParams(iframeStyle)
+  const legacyIframeDimensionParams = getLegacyIframeDimensionParams(rootStyle)
   const forcedOrderDeadline = getForcedOrderDeadline({ swapDeadline, limitDeadline, advancedDeadline })
 
   return {
@@ -252,7 +253,7 @@ function buildWidgetParams(configuratorState: ConfiguratorState | null): CowSwap
     // Layout:
 
     disableScrollbars,
-    iframeStyle,
+    rootStyle,
     bodyWrapperStyle,
     cardStyle,
 
@@ -260,6 +261,7 @@ function buildWidgetParams(configuratorState: ConfiguratorState | null): CowSwap
 
     disableToastMessages,
     disableProgressBar,
+    disableSwitchingTokens,
     disablePostTradeTips,
     disableTokenImport,
     hideRecentTokens,