+++ libs/analytics/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [3.5.2](https://github.com/cowprotocol/cowswap/compare/analytics-v3.5.1...analytics-v3.5.2) (2026-06-25)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-hooks@3.5.0
+  - @cowprotocol/common-utils@3.7.0
+
 ## [3.5.1](https://github.com/cowprotocol/cowswap/compare/analytics-v3.5.0...analytics-v3.5.1) (2026-06-23)
 
 ### Patch Changes