+++ libs/widget-lib/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [4.4.0](https://github.com/cowprotocol/cowswap/compare/widget-lib-v4.3.1...widget-lib-v4.4.0) (2026-06-25)
+
+### Minor Changes
+
+- feat(widget): control tokens switching (#7717)
+
+- feat: Widget Configurator Revamp (#7604)
+
+### Patch Changes
+
+- fix(widget): add iframe loading retry logic (#7636)
+
 ## [4.3.1](https://github.com/cowprotocol/cowswap/compare/widget-lib-v4.3.0...widget-lib-v4.3.1) (2026-06-23)
 
 ### Patch Changes